Make bootstrap_token filter plugin compatible with collections
